feat: replaced the select component
